Add -fPIC, that seems to be the only option to use a static lib in a plugin on 64Bit platforms
CCing kde-buildsystem because I'm not sure this is the right thing to do.
Making it a shared library is a problem because parts of the code is generated
and the generator doesn't allow to add EXPORT macro's. 

The question is: Use -fPIC or build a shared lib that has visibility=default
instead of hidden?

CCMAIL:kde-buildsystem@kde.org


svn path=/trunk/playground/devtools/kdevelop4-extra-plugins/python/; revision=684033