Fixed context ranges
As the python parser does not provide "end" information, we need to go
up the parent chain when encountering, for example, a statement in a
function definition, and adjust the ending positions of all parent nodes
(especially the function definition). Otherwise, the context cannot be
created correctly. This has been a problem before, but reappeared when
we switched to the new python parser. It should be gone now, anyway,
together with the ugly "body.last()", "body.first()" hacks (the
FunctionDefinitionAst node now has proper start and end attribtutes
set).