Added automatic parsing of imported modules
For this, an override for "document()" had to be introduced, because for
some reason, the UseBuilder does not have m_url set (clear from how it
is called, but not clear why this is the case). Sadly, all that stuff is
private, so I just created an extra property called
currentlyParsedDocument. Ugly, but it works.