Fix imports like "import a.b.c".
It just creates a declaration named "a.b.c" now, and the expression
visitor checks whether such a declaration exists when parsing
attributes.
This is very convenient for autocompletion, too, as you are offered the
fully-qualified dotted name instead of the individual components (which
do not make sense at all).

BUG:275486