Replaced previous type hints with HintedType without breaking stuff
no added functionality yet, though.
The idea is this: If a call of a function is encountered, and the types of the call arguments are known,
add a HintedType type to the list of possible types for the argument. This type stores the context it was
created by, and the modification revision it was created in. If read, it should check if the modification
revision matches the current one, and self destruct if it does not.