Fixed some nasty expressionvisitor screwup with return types
The m_lastAccessedReturnType variable must be a stack which contains all return types
which are encountered while visiting a chain of attribute ASTs, like foo().bar().baz(bang()).
This stack is then replayed by the visitCall function. Only by doing this the final result
will have the correct type; the one of the outmost function call which was encountered last in the chain.