Rather hackish fix for declarations being cleaned incorrectly
If there's something like this:
class my_cls():
    def __init__(self):
        self.classattr = 3
        local_var = 12
the declaration for local_var would be cleaned when parsing self.classattr. This
is because self.classattr must be in the class context, and to achieve that the
function context would be closed temporarily, which would clean all declarations
not encountered so far. This is now solved; when temporarily closing contexts, cleaning
is disabled by setting compilingContexts to false for a few lines. A cleaner solution
would be good, tough.