Fixed a bug with argument type hints
The expression visitor needs to be refactored thoroughly. There's quite a
few things there of which I'm not sure how they actually work, and that's not good.
It's getting a bit better with every bug I fix there, because I try replacing that
stuff, but refactoring it completely would probably fix quite a few not-yet-discovered bugs
and make it more maintainable at the same time.