fix a crash bug when renaming classes
I don't like the fix, as I don't know why it works. Apparently dec->inSymbolTable()
is true, but the item isn't really in the symbol table while creating the declaration
after the class has been renamed. It can be worked around by unsetting inSymbolTable
while creating the declaration and setting it again later. It might be that this is
in fact a bug in the duchain. Anyway; it works like this.
BUG:288224