Restructured the instance creation algorithm, to avoid name clashes
BUG:292156
Until now, there was a class __kdev_python_list (or so) which represented
a "list" and its methods; and a function "list" which converted its argument
to a list. This was never very good, but worked so far; but it used to work.
However, inheriting from the list class did not work: There was no class
called "list", this was a function instead.
With this change, this problem should be fixed; the decorators for special
functionality (pass-through of argument types, ...) is now read from the class
constructor of the "list" class, the function is gone.