Fix autocompletion with integral types
like "Foo".|
While fixing this I discovered the general algorithm which collects
possible completion items for attribute access is garbage,
because it relies on the declaration from the expression visitor.
This is something that should be overall removed from the expression visitor
and should be done in the autocompletion code instead.