Use ClassFunctionDeclaration for member functions instead of FunctionDeclaration.
This is meant to be done like that in kdevelop, and is expected in the class browser.
This fixes class member functions from not appearing in the class browser widget.
TODO: It's very unelegant, because it copies all the code required for adding
decorators to function declarations. How could that be done in a better way?