Fix a stupid mistake in the previous commit which prevented the location from updating.
This is about it for the debugger, all the major features are roughly implemented.
I'll now fix the remaining instabilities, then further improvement has to wait
until later, or until someone else wants to do it. Language analysis is more
important for this project.

This branch can be merged to master soon-ish.