Properly implement the variable children fetcher.
Fetching children of variables recursively turned out to be somewhat
difficult: The first level is easy (just use dir, or whatever). After that,
if children (of the children) need to be fetched, the object might not have an
easy python expression to access it with. Imagine a dict which has keys which
are not local variables or similar, and are only accessible through keys() -- what
expression to store in C++ to access this object again if needed?

The current solution saves a dict of id(obj) => obj in python, and sends the
object's id() to C++. This works well, the problem is that this dict controls
the lifetime of the referenced objects. Thus, it has to be cleared often in
order to not screw up the client application's memory management.
Currently it's cleared whenever the user requests a command ("next", ...)
or when the locals table is updated. Thus, it *should* never have influence
on the client application's memory management at all (because the references
only exist while the debugger is paused, and are cleared before continuing).
Still, weakrefs would be better; sadly those are not as versatile in python.

TODO: Investigate what could happen that invalidates those IDs, and investigate
how to properly clear them in that case.
Note: The current solution is crash-prone whatever happens, it just *might* display
outdated data if you somehow manage to delete and re-create an object with the same
ID while the debugger is paused (through command-line-fu,...)

This implementation also enforces that no more children can be fetched while the
client application is running. This could even be made configurable (it's a cool feature)
but for now it's safer like this.