Class and function context ranges now end at the next DEDENT token.
They used to end at the last statement, which sucks for multiple
reasons (autocompletion is one of them). However, making them end
at the next DEDENT is somewhat nontrivial, as it requires extra
information which the parser doesn't provide. This change introduces
a class which reads the file contents again, and extracts indent
information from it. This should not have a large impact on
performance, and will make the planned cleanup of the autocompletion
code much easier.