Fix the background parser from providing a duchain unsuitable to the document revision if parsing fails
If parsing failed, the latest valid DUContext was returned until now.
This caused lots of strange issues, like the mouse over ranges being screwed up,
and the context ranges to be off if lines were inserted somewhere since the
last successful parse run.
Now, ParseJob::translateDUChainToRevision is used to adjust the old duchain ranges
to the changes done to the document.