Another attempt to create a ruleset on when to update a declaration and when not to
The DUChain won't allow a FunctionDeclaration to have a non-function type, and
won't allow a non-FunctionDeclaration to have a function type. That's a problem
for code like this:
foo = 3
def foo(): pass
There's a set of ugly workarounds (which got changed in this commit), but a proper fix
is needed urgently.