More reasonable values for the default parameters
I'd like to just display the original piece of text from the document, but this
is not trivial, as retrieving the text-document from KTE is only possible
if it's loaded (which is not the case in general). In the parse job,
just the byte-array version of the content is available, which has no concept
of lines and columns -- thus, the cursor would need to be translated.
I'm too lazy to do that right now, and also it's quite expensive probably.