Omit the argument types in the completion list, except if display a calltip.
The information is mostly useless, because it's rather inaccurate.
In return, it makes the full completion widget really huge for
functions with many parameters (which often appear in python code).
Also, default parameters are now displayed in brackets instead
of with an "= ..." behind them.
The possible types of arguments are still being displayed in the calltips,
where you might actually be interested in that information (as there's
no overloads in python, I don't think there's many use cases where
the information is very valuable in the completion widget
anyways)