Don't call the visitor base class if not necessary.
This fixes
BUG:308278
The bug was not really a infinite loop or so, it was just that
the additional call massively decreased performance (I think
execution time was like O(e^x) instead of O(x) where x is the number
of elements in the tuple).
Probably more performance can be gained by removing similar statements elsewhere!