Adjust the sip generators so they can theoretically parse pykde
This required support for multi-dot class names, which, for some reason,
had not been considered in the initial design.
The remaining docfiles for pyqt and pykde will be added soon.
A remaining problem is a bug in sip which makes it generate invalid XML:
strings in attributes are not being escaped correctly, and it
often generates stuff like <foobar name="QString("test")"/> :(
For now, I'll just sit down and fix the few occurcences where this happens
by hand.
Long term, the bug in sip should be fixed.