Fix a performance issue with multiple calls in a chain
see unit test from previous commit for an example.
BUG:319055
This commit should of course have been done in 1.5!
The cause for the issue was calling the parent visitor in the Expression
Visitor *plus* creating an extra expression visitor. Creating the
extra visitor is bad enough (n**2 compexity) but additionally calling
AstDefaultVisitor::visit... will blow it up to exp(n).