Fix a crash when using a comprehension as default value for a function def
See duchain/tests/pyduchaintest testCrashes:comprehension_as_default_crash
BUG:321565

The issue was that default values were not taken into account for range
calculation at all, since the ASTDefaultVisitor was missing them
when visiting an ArgumentsAst object. Thus, the range of the function
arguments context ended too early, which resulted in the context for the
comprehension not being created properly.