fix a weird possible crash
It can apparently happen that a function's arguments context is
empty (i.e. no local declarations) even though there's a keyword
argument; I'm not sure how. This prevents a crash happening in that case.
Probably the actual issue is that the declaration for the kwarg is not
being created in some case.