Fix a rather large memory leak in the AST
All the AST nodes were allocated in a memory pool until now, and then when
the AST was not needed any more, the whole pool would be deleted. The
problem is that some of the nodes had QList or QString members, which
allocate memory themselves -- outside of the pool. So the actual AST would
get deleted, but not the memory allocated by QList internally.

It seems to be non-trivial to find a memory pool which can contain a
polymorphic tree of objects and delete them properly (boost::pool can't
do it, for example), and also the usefulness of such a pool becomes
more questionable in that case.
Thus, the pool is now completely gone, and all AST nodes are directly
allocated on the heap. There's a visitor which deletes the AST recursively.

If this works somewhat (which seems to be the case), the AST structure
should be revised to avoid a few allocations. Especially, Identifiers
should be allocated on the stack (they are just 4 bytes and not
polymorphic). It might also make sense to use QVarLengthArray instead of
QList for some properties; the amount of cases where there's more than one
Exception handler is really small.

This change should be backported to 1.6 and eventually even 1.5 if it
proves to be stable enough in master.