Keep the PythonInterp Version in Sync with the PythonLibs Version
Failing to match these versions causes the following problem when
multiple versions of Python 3 are installed: PythonInterp will find the
smalles possible version (e.g. 3.3) and PythonLibs will not try finding
a new one, but will complain instead that 3.4 was not found eve if it is
installed.

Reviewd By: scummos <mail@svenbrauch.de>