fix hinted type appearing repeatedly in unsure type list
The check looking whether a type is already in the list of possible
types for an unsure does not work properly for HintedTypes, because
they create a new IndexedType each (due to different modification
revisions).
When a function declaration was imported from another file, this would
leed to arbitrarily large type lists, since on each parse pass of the
importer an entry would be added but the list was never cleared.

BUG:358563