Restore compile-time support for Python 3.5
Python 3.5 is default on Ubuntu 16.04 LTS and 16.10, Mint 18 and
 other commonly-used distros; we probably shouldn't drop it yet.

AST nodes in 3.6 seem to be a superset of those in 3.5, and no other
 changes (discovered so far?) affect kdev-python. Simply skipping the
 generating code for new AST types will allow compilation against 3.5.

Add a new `SINCE` directive in python36.sdef, e.g. `SINCE 3.6`, and
 make conversionGenerator.py emit #if checks for affected statements.

This can't be used with `CODE`, because that wasn't needed yet. (:P)

The bundled `FindPython` CMake module doesn't handle multiple Python
 versions. Instead use `FindPythonInterp` and `FindPythonLibs` from
 upstream CMake, which (without EXACT) will find newer versions than the
 one requested; the existing check ensures the version isn't too new.

Put #if statements around the new tests, too.

Differential Revision: https://phabricator.kde.org/D3804