Remove unneeded null checks.
In `reopenFittingDeclaration()`, we know the contents of `declarations`
 don't include any null values (all the way from
 `findLocalDeclarationsInternal()`).

In `visitSubscript()`, `->slice` and `->slice->value` are initialised
when the AST is generated, they can't be null if the parser succeeded.

In `findNext()`, we call it only from `visitAttribute()`, where `node`
 is always true. Subsequently, values are checked in the loop before
 assigning them to `parent`.

Above found by Nicolás with clang-analyzer. Notably, each was
 dereferenced without a check elsewhere, so we'd have had crashes if they
 really could be null somehow.

The three AstBuilder::visitFoo() ones  were added by me this morning for
 no reason whatsoever while trying to set a buggiest-patch record.