Remove buggy searching of argument context.
existingDeclarationsForNode() would explicitly search in
 m_mostRecentArgumentsContext whenever that was set, i.e. anywhere
 within a function definition.

In the event of nested function definitions this was incorrect, because
 the outer function's argument context was searched when creating
 argument declarations for the inner function. If an argument name was
 shared by both functions, no declaration would be created for the
 inner function.

This was most noticeable with nested method definitions, because the
 name `self` is almost always shared and the `deprecated name for self`
 warning is wrongly triggered as a side-effect.

DUContext::findDeclarationsInternal() searches imported contexts anyway,
 regardless of the 'DontSearchInParent' flag:

ducontext.h:118 "...(which does not include imported parent-contexts)"

So this code is unnecessary, and can be removed.

One new test fail=>pass.
Fixes one existing test.
No test regressions.