kdevpdb: IBreakpointController API stubs
Breakpoint::deleted() is going away. Same for other obsolete APIs, so
replace the implementation with a stub. The BreakpointController
implementation will be provided at a later point.

This commit begins a major change-set in using the new back-end
kdevpdb.py for the debugger plugin and eventually converting to the
out-of-band communication with the new back-end. Until this development
process is completed, the debugger, or parts of it, are unusable.