kdevpdb: BreakpointController implementation
Implement insertion and removal of breakpoints:
* Inserting and erasing a breakpoint in PausedState and ActiveState now
  works. If a breakpoint is inserted/erased in ActiveState, the
  commands to do so are processed immediately with the help of
  flushCommands().
* Updating the breakpoint's location now works, so breakpoints can be
  added during the debug session.
* Breakpoint enable/disable state is not yet implemented.

The BreakpointController now maintains a QSharedPointer of
BreakpointData for each model breakpoint in a std::vector.

The addBreakpoint() queues 'break' commands only on session startup,
and afterwards this code path is untaken. Just created model
breakpoints don't have a valid location, so after the startup phase,
further PDB breakpoint insertions are handled solely by the
updateBreakpoint() as location updates.

If a queued 'break' command fails: The rejected breakpoint PDB id is
reset, the m_debuggerBreakpointId is decremented, and all greater than
the failed id are decremented so that the PDB breakpoint id
reservations will match the server state.

Breakpoint commands have to be conditionally scheduled in the
updateBreakpoint() or removeBreakpoint() methods since many (possibly
conflicting or would fail) operations can be currently on-going for a
breakpoint. Therefore, the decisions that depend on the breakpointId or
the modelBreakpoint are made in a pass command response handler, where
all of the earlier operations have finished, and so the BreakpointData
is up-to-date.

In general, the deferred handlers must check the modelBreakpoint isn't
a nullptr, because removeBreakpoint() call can take place before the
queued handlers actually run. As an example, removeBreakpoint() can
prevent a 'break' command from being queued in the updateHandler().
There exists other such interactions between the response handlers.