fix python string -> QString conversion
How did this ever work before? PyUnicode_GET_LENGTH is documented
to return the size of the string in code points. Except for UCS4
(which I'm pretty sure was never used... or was it?) this isn't
equal to the length of the array passed to the QString constructor.
This needs to be the size in bytes, which is pretty much completely
unrelated to the amount of code points.

I guess it worked by chance since it typically took the UCS2 path,
and typically all code points fit into 1 UCS2 surrogate. It seems
to use UTF8 now when used from KDevelop (not sure why) and that
makes it completely break down.