kdevpdb: out-of-band communication using a socket
kdevpdbconn.py: unix stream socket helper module.

Implement receiving and sending frames of data through a unix socket
with the kdevpdbconn module. The "protocol" for this is dead simple:

* [4-bytes] A signed 32-bit int. This either the JSON data frame size,
  or if less than one, it works as an escape hatch to encode
  alternative operations, such as a notification to close the
  connection Cmd.Terminate (-1) or an empty data-frame Cmd.DoNothing
  (0).
* [bytes] the frame size worth of raw bytes. If the size is less than
  one, there is no data, and the next frame header immediately follows.

The kdevPdbConnection's  __init__() and establish() are split such that
it's possible to insert an hard-coded frame in self.dataframes before
the connection is established.

Notably, the receiver thread cannot be feasibly eliminated, as it is
required to make progress if both peers send data simultaneously. Thus,
the results of the private __recvDataFrame() are queued until
getDataFrame() extracts the data-frames from the queue.