kdevpdb: main kdevPdb class and out-of-band communication
Re-implement the kdevPdb class utilizing the kdevDbgCore and
kdevpdbconnection modules provided by the previous three commits. Since
kdevPdb is no longer based on Pdb class, the pdb part of the name here
on refers to just Python DeBugger.

The debugger is now nearly silent in respect to stdout/stderr; uncaught
exceptions are printed into stderr, and other messages are outputted via
self.message() and self.error(). However, the debugger remains
completely *broken* until the new out-of-band system is also implemented
on the client plugin side.

On launching kdevpdb.py:
* The debugger accepts "-s <path>" at where to create the socket.
  A "socket:" line is printed to stdout once the debugger is listening
  on the socket. This is required for reliably connecting to the socket
  from the parent process.
* The target command line shall be specified after '--' separator.

The command input, i.e., a request to the debugger, happens via
{"seq":<int>, "input": [<command-string>, <arg>]} data-frames. Omitting
the "input" field from a request allows sending a no-op request to the
debugger, to which the debugger responds with the same seqnro.

Otherwise, the input must be an array, where the first element is a
string specifying the method to be called. The rest of the array
elements (if any) are passed to the method. It is a fatal error for any
exception to escape from the 'do_' named methods, as this will cause the
kdevpdb.py process to print a traceback and exit gracefully.

Resuming commands return from the kdevPdb.enter_debugger() method to
continue running the inferior. Then, the next time enter_debugger() is
called, a "frames" response is appended to report where the debugger
stopped. The response to the last executed command is then sent just
before the next request is acquired with get_next_request().

Responses are sent from the debugger as {"seq":<int>, "data":
[{}]...] }" data-frames. The response "data" value is always a list of
objects. This allows appending any number of objects into a response.
Each such response field provides a string key, which is used to tell
what kind of object it is. Currently, the possible response fields are:

* {"message": "string"} a unstructured output message string.
  This field is added by self.message() and provides all output
  that the may debugger log.
* {"error": "string"} a unstructured error message string. This field is
  added by self.error() in cases where a command has failed.
* {"frames": [ <list-of-stack-frames> ]}
* {"activeframe": int()}
* {"breakpoints": [{<breakpoint-info>}]}
  This field reports information about breakpoints.

More of such response fields can be defined in later commits to
implement more functionality.

Although the client can fully determine what seqnro is used through the
request-response round-trip, it shall use a max(last-request-seqnro +
1, last-response-seqnro + 1) as the new seqnro when sending a request.
This because the kdevpdb.py will send, on start up, a response which
has no client sent request associated with it.