kdevpdb: communicate using kdevPdbConnection
Replace the in-band communication through stdout with an out-of-band
system using kdevPdbConnection. The debugger is now nearly silent in
respect to stdout/stderr; uncaught exceptions are still printed as
normal. However, the debugger remains completely *broken* until the new
communication system is also implemented on the debugger plugin side.

The debugger main() function is implemented in a new kdevpdbmain.py
source file, which then imports the kdevpdb module as a local. This
simplifies using modules from kdevpdb.py, which would otherwise get
dropped due to the namespace replacing in the debugger._run().

On launching kdevpdbmain.py:

* The debugger accepts "-s <path>" at where to create the socket.
  A "socket:" line is printed to stdout once the debugger is listening
  on the socket. This is required for reliably connecting to the socket
  from the parent process.
* The target command line shall be specified after '--' separator.

The command input, i.e., a request to the debugger, happens with
{"seq":<int>, "input":<command-string> } data-frames. The
<command-string> must be terminated by a '\n' char. The "seq" field in
the requests and responses enables keeping track of the queued
requests' responses.

Certain commands cause kdevPdb.cmdloop() to return; therefore, the
response is only sent after a request has been fully completed just
before acquiring the next request with getDataFrame().

Now, responses are sent from the debugger as {"seq":<int>, "data":
[{}]...] }" data-frames. The response "data" value is always a list of
objects. This allows appending any number of objects into a response.
Each such response field provides a string key, which is used to tell
what kind of object it is. Currently, the possible response fields are:

* {"message": "string"} a unstructured output message string.
  This field is added by self.message() and provides all output which
  is otherwise generated by Pdb.
* {"error": "string"} a unstructured error message string.
  This field is added by self.error() and provides all errors which are
  otherwise generated by Pdb.
* {"frames": [ <list-of-stack-frames> ]}

More of such response fields will be defined in later commits. It is
preferred to implement our own versions of used commands to structure
them instead of relying on the unstructured "message" fields from
the Pdb and Bdb module.

Although the client can fully determine what seqnro is used through the
request-response round-trip, it shall use a max(last-request-seqnro +
1, last-response-seqnro + 1) as the new seqnro when sending a request.
This because the kdevpdb.py will send, on start up, a response which
has no client sent request associated with it.