kdevpdb: PdbProcess implementation
The server/inferior can now be managed with a PdbProcess, which handles
the I/O, creating and terminating of the kdevpdbmain.py debugger
process.

Processing of commands is excluded when the inferior's code is being
executed. tryInterrupt() can revive the processing of commands from a
such state. (e.g. inferior is spinning in a infinite loop.)

However, any sent ControlCommands would be immediately processed by the
kdevpdb.py's receiver thread. Because of this, PdbProcess waits for an
equal number of data frames to be received as was sent before it sends
a ControlCommand.

The debugger's UI needs a way to tell when the debugger is busy
executing the inferior's code or processing commands and when it should
attempt to update all its widgets. PdbProcess::running() and
PdbProcess::suspended() signals are provided to do such tasks.

The reading of '~/.pdbrc' or '.pdbrc' files is disabled, as this can
possibly break or interfere the initialization of PdbProcess.