kdevpdb: PdbProcess implementation
The server/inferior can now be managed with a PdbProcess, which handles
the I/O, creating and terminating of the kdevpdb.py debugger process.

Processing of commands is excluded when the inferior's code is being
executed. tryInterrupt() can revive the processing of commands from a
such state. (e.g. inferior is spinning in a infinite loop.)

As any sent ControlCommands would be immediately processed by the
kdevpdb.py's receiver thread, PdbProcess waits for an equal number of
data frames to be received before it sends a ControlCommand.

The debugger's UI needs a way to tell when the debugger is busy
executing the inferior's code or processing commands. The
PdbProcess::running() and PdbProcess::suspended() signals are provided
for this purpose.

v2: Adapt to the recently introduced StartupInfo and update the code
    documentation a bit.