kdevpdb: improve the interruption of commands
Pdb does not handle interrupting a 'return' command well. If we have
only one inferior stack-frame, interrupting a 'return' in this scenario
results in the inferior observing an KeyboardInterrupt exception, and
thus causing the Pdb to wrongly enter post-mortem debugging.

The fix is to override Pdb's SIGINT handler and install our own, which
now handles this quirk of Pdb and never raises an KeyboardInterrupt.

Also, structure the interrupt handler response such that we can tell
this apart from a un-interrupted response of an command.