Fix a crash in CodeHelpers::expressionUnderCursor()
The document line can be empty, resulting in a crash or a failed
assert.

Return an empty string if the document line is empty; otherwise, limit
the index to inside the document line. Add also a test for the
expressionUnderCursor() for code-coverage mostly. This function is
critical in parsing the document's source code in a debug session and
in code completion.