kdevpdb: PdbDebuggerInstance higher-level requests v2
PdbProcess manages the communication with the kdevpdbmain.py, but it
knows nothing about the content of the requests or responses.
Therefore, implement a PdbDebuggerInstance class that handles the
higher-level requests and the invocation of response handlers. This
also hides the lower-level class for use-cases, which only need the
PdbDebuggerInstance.

The command sequence number (seqnro) handling is mostly abstracted
away, so the class user doesn't need to care about this detail anymore.
The initial response "seq":-1 that the kdevpdb.py sends once up on
start-up, can be processed by assigning a handler for seqnro -1 with
setMethod() just before creating the process.

The PdbDebuggerInstance::request() can accept any Pdb command to be
run. Importantly, a request() call is non-blocking, so it can be safely
called from the main thread. The request() also allows registering an
response callback to be invoked when the command's response is
available from the debugger.

The data-frames received from PdbProcess are minimally decoded and
queued up. Later, the deferred process() picks an JSON response from
the queue and splits it into a list of response fields. A response
handler is expected to pick its needed data from this list, and a few
helpers are provided for this task.

Each time the response queue is emptied, PdbProcess::signalSuspended()
is called, which can emit PdbProcess::suspended() if there are no
ongoing requests.