kdevpdb: Rewrite of DebugSession core
This commit re-implements the core of the debugger such that:
* Most of the user actions for stepping, continuing and like do work.
* Interrupting a long running "continue" or "return" command works;
  This is an major improvement compared to the previous situation.
* Proper ActiveState/PausedState switching; The debugger only enters
  ActiveState, when the inferior's code is going to be executed.
* Frame-stack view works.
* Execution location update works.
* Inferior's stderr/stdout is simply forwarded into the console as is.
* Killing the session/debugger works.

Possible debugger hangs are resolved by launching a timer in the
stopDebugger() action to ensure the debugger process will be killed
after 5 seconds. Normally, the process ends fast enough before this, so
the timer won't expire.

Everything else should be assumed to be broken or missing; No variable
support is implemented at all, selecting a stack-frame does nothing,
and breakpoints are neither implemented at all. Subsequent commits will
reintroduce more of the now lost functionality a case by case basis.